cloud_functions/create_signed_url/requirements.txt
google-cloud-storage
firebase-admin